+++ PrototypeWithAuth/Controllers/ApplicationUsersController.cs
@@ -170,7 +170,7 @@ namespace PrototypeWithAuth.Controllers
                 Employees = workerHoursViewModel,
                 Months = Enumerable.Range(1, 12).ToList(),
                 Years = Enumerable.Range(2010, DateTime.Today.Year - 2009).ToList(),
-                TotalWorkingDaysInMonth = AppUtility.GetTotalWorkingDaysThisMonth(new DateTime(DateTime.Now.Year, month, 1), _context.CompanyDayOffs, 0),
+                TotalWorkingDaysInMonth = (int)AppUtility.GetTotalWorkingDaysThisMonth(new DateTime(DateTime.Now.Year, month, 1), _context.CompanyDayOffs, 0),
                 TotalWorkingDaysInYear = amountInYear == 0 ? AppUtility.GetTotalWorkingDaysThisYear(new DateTime(year, 1, 1), _context.CompanyDayOffs, 0) : amountInYear
             };
             return viewModel;
